# specs/001-stock-portfolio-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MyStock API
  description: |
    RESTful API for MyStock stock portfolio management application.
    
    ## Features
    - User authentication with JWT tokens
    - Watchlist management with drag-and-drop ordering
    - Portfolio tracking across three categories (장기/단기/정찰병)
    - Real-time stock data integration with Alpha Vantage API
    - Dark mode preference persistence
    
    ## Authentication
    Most endpoints require authentication via JWT bearer token.
    Obtain token via `/api/v1/auth/login` endpoint.
    
  version: 1.0.0
  contact:
    name: MyStock API Support
    email: support@mystock3.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mystock3.example.com/api/v1
    description: Production server
  - url: https://api-staging.mystock3.example.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /stocks/search:
    get:
      tags:
        - stocks
      summary: Search stocks
      description: Search for stocks by symbol or company name
      operationId: searchStocks
      parameters:
        - name: q
          in: query
          description: Search query (symbol or company name)
          required: true
          schema:
            type: string
            minLength: 1
            example: "AAPL"
        - name: limit
          in: query
          description: Maximum number of results
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockSearchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /stocks/{symbol}/quote:
    get:
      tags:
        - stocks
      summary: Get stock quote
      description: |
        Retrieve current stock quote data. 
        Data is cached for 1 minute to minimize external API calls.
      operationId: getStockQuote
      parameters:
        - name: symbol
          in: path
          description: Stock ticker symbol
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
            example: "AAPL"
      responses:
        '200':
          description: Stock quote retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockQuote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Stock symbol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "종목을 찾을 수 없음"
                error_code: "STOCK_NOT_FOUND"
                timestamp: "2025-11-05T10:00:00Z"
        '429':
          description: API rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "API rate limit exceeded, please retry later"
                error_code: "RATE_LIMIT_EXCEEDED"
                timestamp: "2025-11-05T10:00:00Z"
        '503':
          description: External stock API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Stock data service temporarily unavailable"
                error_code: "EXTERNAL_API_ERROR"
                timestamp: "2025-11-05T10:00:00Z"
  /stocks/{symbol}/history:
    get:
      tags:
        - stocks
      summary: Get historical stock data
      description: Retrieve historical price data for charting (candlestick)
      operationId: getStockHistory
      parameters:
        - name: symbol
          in: path
          description: Stock ticker symbol
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
        - name: period
          in: query
          description: Time period for historical data
          required: false
          schema:
            type: string
            enum: [1D, 5D, 1M, 3M, 6M, 1Y, 5Y]
            default: 1M
      responses:
        '200':
          description: Historical data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint (7-day expiration)
...
  schemas:
...
    StockSearchResult:
      type: object
      required:
        - symbol
        - name
      properties:
        symbol:
          type: string
          example: "AAPL"
        name:
          type: string
          example: "Apple Inc."
        type:
          type: string
          example: "Common Stock"
        region:
          type: string
          example: "United States"
        currency:
          type: string
          example: "USD"
    StockQuote:
      type: object
      required:
        - symbol
        - company_name
        - current_price
        - change
        - change_percent
        - open
        - high
        - low
        - volume
        - last_updated
        - currency
      properties:
        symbol:
          type: string
          example: "AAPL"
        company_name:
          type: string
          example: "Apple Inc."
        current_price:
          type: number
          format: double
          example: 178.32
        change:
          type: number
          format: double
          description: Price change from previous close
          example: 2.15
        change_percent:
          type: number
          format: double
          description: Percentage change
          example: 1.22
        open:
          type: number
          format: double
          example: 176.50
        high:
          type: number
          format: double
          example: 179.00
        low:
          type: number
          format: double
          example: 176.20
        volume:
          type: integer
          format: int64
          example: 58392010
        last_updated:
          type: string
          format: date-time
          description: API data timestamp
          example: "2025-11-05T11:30:00Z"
        currency:
          type: string
          example: "USD"
    StockHistoryResponse:
      type: object
      required:
        - symbol
        - period
        - data
      properties:
        symbol:
          type: string
          example: "AAPL"
        period:
          type: string
          example: "1M"
        data:
          type: array
          items:
            $ref: '#/components/schemas/StockHistoryPoint'
    StockHistoryPoint:
      type: object
      required:
        - date
        - open
        - high
        - low
        - close
        - volume
      properties:
        date:
          type: string
          format: date
          example: "2025-11-05"
        open:
          type: number
          format: double
          example: 176.50
        high:
          type: number
          format: double
          example: 179.00
        low:
          type: number
          format: double
          example: 176.20
        close:
          type: number
          format: double
          example: 178.32
        volume:
          type: integer
          format: int64
          example: 58392010

    # Error Schemas
    ErrorResponse:
      type: object
      required:
        - detail
        - error_code
        - timestamp
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"
        error_code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp
          example: "2025-11-05T10:00:00Z"
        field_errors:
          type: object
          description: Field-specific validation errors
          additionalProperties:
            type: string
          example:
            email: "Invalid email format"
            password: "Password must be at least 6 characters"
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Validation error"
            error_code: "VALIDATION_ERROR"
            timestamp: "2025-11-05T10:00:00Z"
            field_errors:
              symbol: "Symbol must be uppercase letters only"
    Unauthorized:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Authentication required"
            error_code: "UNAUTHORIZED"
            timestamp: "2025-11-05T10:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Resource not found"
            error_code: "NOT_FOUND"
            timestamp: "2025-11-05T10:00:00Z"